Move PSO example
