+++ example/src/main.rs
@@ -1,5 +1,6 @@
 #![allow(dead_code)]
 mod ga;
+mod pso;
 
 fn main() {
 	ga::ga_example();
